bugfix: make `make create-event` work consistently
Change the Accept header for cheerio.fromURL to `*/*`.

By default, cheerio.fromURL runs with this Accept header:

    `'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8'`

For some reason, this results in Meetup.com returning content inconsistently
when running `make create-event`. To test out our hypothesis that the Accept
header is the culprit, run the following curl commands:

    $ for i in {1..10}; do curl -sL -H "Accept: text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8" https://www.meetup.com/aws-sg/events/310825790 | wc -c; done;
    93441
    63083
    93441
    93441
    93441
    63083
    93441
    93441
    63083
    63083

    $ for i in {1..10}; do curl -sL -H "Accept: */*" https://www.meetup.com/aws-sg/events/310825790 | wc -c; done;
    63084
    63084
    63084
    63084
    63084
    63084
    63084
    63084
    63084
    63084

It is hard to determine exactly why this happens. For now, fix it by using */*.

Consequently, this change may reduce the reliability of the scraper in general
since some sites may flag `*/*` as non-human. We may want to have a per-site
list of Accept headers to use in the future.

Misc: some tabs -> spaces thing.